zed: hide Install CLI menu button on Windows
