+++ crates/install_cli/src/register_zed_scheme.rs
@@ -1,0 +1,15 @@
+use client::ZED_URL_SCHEME;
+use gpui::{AsyncApp, actions};
+
+actions!(
+    cli,
+    [
+        /// Registers the zed:// URL scheme handler.
+        RegisterZedScheme
+    ]
+);
+
+pub async fn register_zed_scheme(cx: &AsyncApp) -> anyhow::Result<()> {
+    cx.update(|cx| cx.register_url_scheme(ZED_URL_SCHEME))?
+        .await
+}
